version 0.6 - March 2011
1. Improved the intersect tool.  Now SNPs and MNPs are handled (although
work still needs to be done on variants with multiple alternates).  If SNPs
are found at the same locus in two different files with different alternate
alleles, both records are written out.  The two SNPs are different and will
not be included in the output if common SNPs are requested.  SNPs with
multiple alternates are also treated as distinct from a SNP with a sinlge
alternate even if that SNP shares a common alternate.  The assumption is
that A G/T is a SNP with evidence for both G and T alternates, whereas
A G does not show evidence for T as an alternate and is, as such, a
different SNP.