version 0.7 - March 2011
1. Modified the annotate tool.  The tool now accepts dbSNP and hapmap
vcf files as well as bed files.  The bed file must have the regular
reference sequence and start and end coordinates.  The text in the
fourth column will be appended to the info string in the vcf record.
Bed files should not include overlapping regions.  The tool has also
been modified to allow the user to specify any of the above options
simultaneously (although, currently, only one bed file at a time).