version 0.9 - April 2011
1. Added --delete-info, -d to the filter tool.  This allows the user to remove
this info ID from the info field in all records.

2. Added --fail-filter, -f to the filter tool.  This allows all variants that do
not have "PASS" in the filter line to be filtered out.

3. Added --strip-records, -s to the filter tool.  This accepts a comma separated
list of info IDs and any variant containing any of these will be filtered out.
Logic has not been added to this as yet (e.g. strip record if X and Y etc).

4. Added --pass-filter, -p to the intersect tool.  If this flag is added, only
variants marked as "PASS" will be considered when performing intersections.