version 0.11 - April 2011
1. Major overhaul of the logic.  When parsing a vcf file, the reords
are loaded into a map and then the map is looped and interated.  This
ensures that if the vcf file has some entries out of order, they will
be output in order.  Each variant in a multiple alternate format is
stored as a separate variant at its location.  Not all tools have been
modified as yet.  Functioning are: intersect, stats and filter.

2. Added --find-hets, -e to filter tool.  If genotype information is
present, the genotypes are parsed to find all hets and the sample
names are added to the info string as HET=SAMPLE.