version 0.13 - April 2011
1. Updated the bed intersection tool.  The bed file is treated
similarly to the vcf files with each interval added to a
structure.  This ensures proper sorting of the intervals.  The
tool also checks to see if intervals overlap and if so, breaks them
up into unique intervals, carrying across any annotations that
were associated with each bed record (the annotation tool still
requires updating).