version 0.34 - June 2011
1. Added the --split-mnps (-p) option to the filter tool.  This will
take all MNPs present in the vcf file and decompose them into a set
of SNPs.  These SNPs are tagged with the FROM_MNP flag.  Loci are
subsequently checked to ensure that there is only a single SNP with
same alternate allele present at the locus.  If there are duplicates,
the SNP from the MNP is discarded.  For example, if an A/C SNP is
present at position 20, as is the MNP AG/CT, decomposing the MNP will
generate a new G/T SNP at position 21, but will generate an A/C SNP at
position 20.  This SNP is already present, so the decomposition only
contributes a new SNP at position 21.