version 0.43 - September 2011
1. Modified structures so that multiple records at the same locus
will be read in and used in all tools.  For output, there is no
attempt to merge these records into a single record.  It is
possible that these variant observations could be from separate
sources and merging them together would lose information (for
example, the depths may be different, but this is a single value
only).

2. Fixed a bug in the intersection tool.  The final record in a
vcf file could be treated incorrectly.